requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart
google-genai>=1.0.0
sqlalchemy==2.0.36
sqlalchemy-cockroachdb==2.0.2
psycopg2-binary==2.9.10
pydantic==2.9.2
python-dotenv
mangum
#passlib[bcrypt]
passlib
streamlit
requests
pandas
thefuzz             #<-- NEW: For text matching
python-levenshtein  #<-- NEW: Makes fuzz faster
python-dateutil     #<-- NEW: For smart date parsing
altair              #<-- NEW: For charts with labels